Refactor files, added test for platinum plan for elasticsearch
