+++ internal/providers/terraform/ibm/testdata/database_test/database_test.tf
@@ -94,3 +94,40 @@ resource "ibm_database" "test_es_enterprise_db2" {
     }
   }
 }
+
+resource "ibm_database" "test_es_platinum_db1" {
+  name                         = "demo-es-platinum"
+  service                      = "databases-for-elasticsearch"
+  plan                         = "platinum"
+  location                     = "eu-gb"
+
+  group {
+    group_id = "member"
+    memory {
+      allocation_mb = 12288
+    }
+    disk {
+      allocation_mb = 131072
+    }
+    cpu {
+      allocation_count = 3
+    }
+  }
+}
+
+resource "ibm_database" "test_es_platinum_db2" {
+  name                         = "demo-es-platinum2"
+  service                      = "databases-for-elasticsearch"
+  plan                         = "platinum"
+  location                     = "eu-gb"
+
+  group {
+    group_id = "member"
+    members {
+      allocation_count = 4
+    }
+    memory {
+      allocation_mb = 1024
+    }
+  }
+}
