+++ internal/providers/terraform/ibm/testdata/database_test/database_test.usage.yml
@@ -14,3 +14,10 @@ resource_usage:
   ibm_database.test_es_enterprise_db2:
     elasticsearch_database_ram_mb: 15360
     elasticsearch_database_members: 4
+  ibm_database.test_es_platinum_db1:
+    elasticsearch_database_ram_mb: 12288
+    elasticsearch_database_disk_mb: 131072
+    elasticsearch_database_core: 3
+  ibm_database.test_es_platinum_db2:
+    elasticsearch_database_ram_mb: 1024
+    elasticsearch_database_members: 4
